requirements.txt
langchain==0.0.354
openai==1.6.1
chromadb==0.4.22
streamlit==1.29.0
python-dotenv
lark==1.1.8
pandas==2.1.4
tiktoken
unstructured==0.11.8